refactor, add add-skills-to-statistics to statistics-service
