+++ src/utils/utils.ts
@@ -1,4 +1,5 @@
 import { SkillsMap } from "../controllers/statisticsController";
+import Skill from "../models/skill";
 
 export const convertSkillsMapToSkillIdList = (skillsMap: SkillsMap) => {
   const skillIdList: string[] = [];
@@ -6,4 +7,14 @@ export const convertSkillsMapToSkillIdList = (skillsMap: SkillsMap) => {
     skillIdList.push(skillId);
   });
   return skillIdList;
+};
+
+export const convertSkillListToMap = (skillList: Skill[]) => {
+  const skillsMap: SkillsMap = {};
+
+  skillList.forEach((skill) => {
+    skillsMap[skill._id!.toString()] = skill;
+  });
+
+  return skillsMap;
 };
